test(winglet): split over-cap test records and link specs to contracts
테스트 레코드 32케이스 상한 초과 9파일을 describe 주제 단위로 분할(케이스 무변경·합계 보존):
shallowClone 40=28+12 · stableEquals 43=25+8+10 · MessageChannelScheduler 33=27+6 ·
compare.strict 34=19+15 · compare 46=31+15 · difference 37=20+17 ·
isCompatibleSchemaType 96=30+32+23+11 · isIdenticalSchemaType 97=31+22+32+12 ·
schemaTypeFilters 43=28+9+6.
json-schema spec 2파일에 계약 그룹 링크 마커 추가(src DETAIL의 scanner-sync/scanner-async
그룹과 연결).

검증: 케이스 계수 전 조각 ≤32·exact, common-utils 1105·json 570(#333과 동수)·
json-schema 392·react-utils 183 통과
