+++ packages/winglet/common-utils/src/utils/scheduler/MessageChannelScheduler/__tests__/MessageChannelScheduler.performance.test.ts
@@ -1,0 +1,381 @@
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+
+import { MessageChannelScheduler } from '../MessageChannelScheduler';
+import { destroyGlobalScheduler, setImmediate } from '../handler';
+
+describe('MessageChannelScheduler', () => {
+  beforeEach(() => {
+    // Reset singleton instance and global scheduler
+    (MessageChannelScheduler as any).__instance__ = null;
+    destroyGlobalScheduler();
+  });
+
+  afterEach(() => {
+    destroyGlobalScheduler();
+  });
+
+  describe('Performance Comparison: MessageChannel vs setTimeout', () => {
+    it('should measure execution latency comparison', async () => {
+      const iterations = 100;
+      const messageChannelTimings: number[] = [];
+      const setTimeoutTimings: number[] = [];
+
+      // Test MessageChannel latency
+      for (let i = 0; i < iterations; i++) {
+        await new Promise<void>((resolve) => {
+          const start = performance.now();
+          setImmediate(() => {
+            messageChannelTimings.push(performance.now() - start);
+            resolve();
+          });
+        });
+      }
+
+      // Test setTimeout(0) latency
+      for (let i = 0; i < iterations; i++) {
+        await new Promise<void>((resolve) => {
+          const start = performance.now();
+          setTimeout(() => {
+            setTimeoutTimings.push(performance.now() - start);
+            resolve();
+          }, 0);
+        });
+      }
+
+      const avgMessageChannel =
+        messageChannelTimings.reduce((a, b) => a + b, 0) / iterations;
+      const avgSetTimeout =
+        setTimeoutTimings.reduce((a, b) => a + b, 0) / iterations;
+
+      console.log(
+        `\n=== Performance Comparison (${iterations} iterations) ===`,
+      );
+      console.log(
+        `MessageChannel avg latency: ${avgMessageChannel.toFixed(3)}ms`,
+      );
+      console.log(`setTimeout(0) avg latency: ${avgSetTimeout.toFixed(3)}ms`);
+      console.log(
+        `Speedup: ${(avgSetTimeout / avgMessageChannel).toFixed(2)}x`,
+      );
+
+      // MessageChannel should generally be faster or comparable
+      expect(avgMessageChannel).toBeLessThanOrEqual(avgSetTimeout * 1.5);
+    });
+
+    it('should compare batch task scheduling performance', async () => {
+      const batchSize = 1000;
+      let messageChannelCompleted = 0;
+      let setTimeoutCompleted = 0;
+
+      // Batch schedule with MessageChannel
+      const mcStart = performance.now();
+      const mcPromise = new Promise<number>((resolve) => {
+        for (let i = 0; i < batchSize; i++) {
+          setImmediate(() => {
+            messageChannelCompleted++;
+            if (messageChannelCompleted === batchSize) {
+              resolve(performance.now() - mcStart);
+            }
+          });
+        }
+      });
+
+      const mcTime = await mcPromise;
+
+      // Batch schedule with setTimeout
+      const stStart = performance.now();
+      const stPromise = new Promise<number>((resolve) => {
+        for (let i = 0; i < batchSize; i++) {
+          setTimeout(() => {
+            setTimeoutCompleted++;
+            if (setTimeoutCompleted === batchSize) {
+              resolve(performance.now() - stStart);
+            }
+          }, 0);
+        }
+      });
+
+      const stTime = await stPromise;
+
+      console.log(`\n=== Batch Scheduling (${batchSize} tasks) ===`);
+      console.log(`MessageChannel total time: ${mcTime.toFixed(2)}ms`);
+      console.log(`setTimeout(0) total time: ${stTime.toFixed(2)}ms`);
+      console.log(`Speedup: ${(stTime / mcTime).toFixed(2)}x`);
+
+      expect(messageChannelCompleted).toBe(batchSize);
+      expect(setTimeoutCompleted).toBe(batchSize);
+    });
+
+    it('should measure task execution order consistency', async () => {
+      const executionOrder: string[] = [];
+
+      // Schedule mixed macro tasks
+      await Promise.resolve().then(() => {
+        executionOrder.push('microtask-1');
+      });
+
+      setImmediate(() => {
+        executionOrder.push('messageChannel-1');
+      });
+
+      setTimeout(() => {
+        executionOrder.push('setTimeout-1');
+      }, 0);
+
+      setImmediate(() => {
+        executionOrder.push('messageChannel-2');
+      });
+
+      setTimeout(() => {
+        executionOrder.push('setTimeout-2');
+      }, 0);
+
+      await Promise.resolve().then(() => {
+        executionOrder.push('microtask-2');
+      });
+
+      // Wait for all tasks
+      await new Promise((resolve) => setTimeout(resolve, 100));
+
+      console.log('\n=== Execution Order ===');
+      console.log(executionOrder);
+
+      // Microtasks should always execute first
+      expect(executionOrder[0]).toBe('microtask-1');
+      expect(executionOrder[1]).toBe('microtask-2');
+
+      // Macro tasks should follow
+      expect(executionOrder.slice(2)).toContain('messageChannel-1');
+      expect(executionOrder.slice(2)).toContain('messageChannel-2');
+      expect(executionOrder.slice(2)).toContain('setTimeout-1');
+      expect(executionOrder.slice(2)).toContain('setTimeout-2');
+    });
+  });
+
+  describe('Browser Environment Performance Comparison', () => {
+    it('should compare MessageChannel vs setTimeout vs scheduleMacrotask in browser-like conditions', async () => {
+      const iterations = 50;
+
+      // Test individual task latency - browser-like conditions
+      const messageChannelTimings: number[] = [];
+      const setTimeoutTimings: number[] = [];
+      const scheduleMacrotaskTimings: number[] = [];
+
+      console.log('\n=== Browser-like Individual Task Latency Test ===');
+
+      // MessageChannel (setImmediate) latency
+      for (let i = 0; i < iterations; i++) {
+        await new Promise<void>((resolve) => {
+          const start = performance.now();
+          setImmediate(() => {
+            messageChannelTimings.push(performance.now() - start);
+            resolve();
+          });
+        });
+        // Add a small delay for browser-like conditions
+        await new Promise((resolve) => setTimeout(resolve, 1));
+      }
+
+      // setTimeout(0) latency
+      for (let i = 0; i < iterations; i++) {
+        await new Promise<void>((resolve) => {
+          const start = performance.now();
+          setTimeout(() => {
+            setTimeoutTimings.push(performance.now() - start);
+            resolve();
+          }, 0);
+        });
+        await new Promise((resolve) => setTimeout(resolve, 1));
+      }
+
+      // scheduleMacrotask latency (depends on the current environment)
+      const { scheduleMacrotask } = await import('../../scheduleMacrotask');
+      for (let i = 0; i < iterations; i++) {
+        await new Promise<void>((resolve) => {
+          const start = performance.now();
+          scheduleMacrotask(() => {
+            scheduleMacrotaskTimings.push(performance.now() - start);
+            resolve();
+          });
+        });
+        await new Promise((resolve) => setTimeout(resolve, 1));
+      }
+
+      const avgMessageChannel =
+        messageChannelTimings.reduce((a, b) => a + b, 0) / iterations;
+      const avgSetTimeout =
+        setTimeoutTimings.reduce((a, b) => a + b, 0) / iterations;
+      const avgScheduleMacrotask =
+        scheduleMacrotaskTimings.reduce((a, b) => a + b, 0) / iterations;
+
+      console.log(
+        `MessageChannel (setImmediate) avg: ${avgMessageChannel.toFixed(3)}ms`,
+      );
+      console.log(`setTimeout(0) avg: ${avgSetTimeout.toFixed(3)}ms`);
+      console.log(
+        `scheduleMacrotask avg: ${avgScheduleMacrotask.toFixed(3)}ms`,
+      );
+      console.log(
+        `MessageChannel speedup over setTimeout: ${(avgSetTimeout / avgMessageChannel).toFixed(2)}x`,
+      );
+      console.log(
+        `MessageChannel speedup over scheduleMacrotask: ${(avgScheduleMacrotask / avgMessageChannel).toFixed(2)}x`,
+      );
+
+      // Verify results — latency figures above are observational only.
+      // This suite runs on Node, where MessageChannel falls back to
+      // setImmediate; its latency relative to setTimeout(0) flips under host
+      // load and GC timing, so asserting an ordering here is not reproducible.
+      // Only measurement completeness is asserted.
+      expect(messageChannelTimings.length).toBe(iterations);
+      expect(setTimeoutTimings.length).toBe(iterations);
+      expect(scheduleMacrotaskTimings.length).toBe(iterations);
+    });
+
+    it('should demonstrate the 4ms setTimeout delay issue vs MessageChannel', async () => {
+      const iterations = 20;
+
+      console.log('\n=== 4ms setTimeout Delay vs MessageChannel Demo ===');
+
+      // Test rapid sequential scheduling to trigger 4ms delay
+      const setTimeoutDelays: number[] = [];
+      const messageChannelDelays: number[] = [];
+
+      // Rapid setTimeout scheduling
+      const setTimeoutStart = performance.now();
+      let setTimeoutCompleted = 0;
+
+      for (let i = 0; i < iterations; i++) {
+        const taskStart = performance.now();
+        setTimeout(() => {
+          setTimeoutDelays.push(performance.now() - taskStart);
+          setTimeoutCompleted++;
+        }, 0);
+      }
+
+      // Wait for setTimeout tasks to complete
+      while (setTimeoutCompleted < iterations) {
+        await new Promise((resolve) => setTimeout(resolve, 10));
+      }
+      const setTimeoutTotal = performance.now() - setTimeoutStart;
+
+      // Rapid MessageChannel scheduling
+      const messageChannelStart = performance.now();
+      let messageChannelCompleted = 0;
+
+      for (let i = 0; i < iterations; i++) {
+        const taskStart = performance.now();
+        setImmediate(() => {
+          messageChannelDelays.push(performance.now() - taskStart);
+          messageChannelCompleted++;
+        });
+      }
+
+      // Wait for MessageChannel tasks to complete
+      while (messageChannelCompleted < iterations) {
+        await new Promise((resolve) => setTimeout(resolve, 5));
+      }
+      const messageChannelTotal = performance.now() - messageChannelStart;
+
+      const avgSetTimeoutDelay =
+        setTimeoutDelays.reduce((a, b) => a + b, 0) / iterations;
+      const avgMessageChannelDelay =
+        messageChannelDelays.reduce((a, b) => a + b, 0) / iterations;
+
+      console.log(`\nRapid Sequential Scheduling (${iterations} tasks):`);
+      console.log(
+        `setTimeout(0) - Total: ${setTimeoutTotal.toFixed(2)}ms, Avg per task: ${avgSetTimeoutDelay.toFixed(3)}ms`,
+      );
+      console.log(
+        `MessageChannel - Total: ${messageChannelTotal.toFixed(2)}ms, Avg per task: ${avgMessageChannelDelay.toFixed(3)}ms`,
+      );
+      console.log(
+        `Total time speedup: ${(setTimeoutTotal / messageChannelTotal).toFixed(2)}x`,
+      );
+      console.log(
+        `Per-task speedup: ${(avgSetTimeoutDelay / avgMessageChannelDelay).toFixed(2)}x`,
+      );
+
+      // Speedup figures above are observational only. Under rapid scheduling on
+      // Node the two schedulers routinely trade places (observed: setImmediate
+      // 1.374ms avg vs setTimeout 0.563ms avg), so no ordering is asserted —
+      // only that every scheduled task actually ran.
+      expect(setTimeoutDelays.length).toBe(iterations);
+      expect(messageChannelDelays.length).toBe(iterations);
+    });
+
+    it('should show execution timing with mixed task types', async () => {
+      const executionOrder: { type: string; timestamp: number }[] = [];
+      const baseTime = performance.now();
+
+      console.log('\n=== Mixed Task Execution Order ===');
+
+      // Schedule various task types
+      Promise.resolve().then(() => {
+        executionOrder.push({
+          type: 'Promise.resolve()',
+          timestamp: performance.now() - baseTime,
+        });
+      });
+
+      queueMicrotask(() => {
+        executionOrder.push({
+          type: 'queueMicrotask',
+          timestamp: performance.now() - baseTime,
+        });
+      });
+
+      setImmediate(() => {
+        executionOrder.push({
+          type: 'setImmediate (MessageChannel)',
+          timestamp: performance.now() - baseTime,
+        });
+      });
+
+      setTimeout(() => {
+        executionOrder.push({
+          type: 'setTimeout(0)',
+          timestamp: performance.now() - baseTime,
+        });
+      }, 0);
+
+      setTimeout(() => {
+        executionOrder.push({
+          type: 'setTimeout(1)',
+          timestamp: performance.now() - baseTime,
+        });
+      }, 1);
+
+      // Wait for all tasks
+      await new Promise((resolve) => setTimeout(resolve, 50));
+
+      console.log('Execution order with timestamps (ms from start):');
+      executionOrder
+        .sort((a, b) => a.timestamp - b.timestamp)
+        .forEach(({ type, timestamp }) => {
+          console.log(`  ${timestamp.toFixed(3)}ms: ${type}`);
+        });
+
+      // Verify microtasks execute before macrotasks
+      const microtaskTasks = executionOrder.filter(
+        (task) =>
+          task.type.includes('Promise') || task.type.includes('queueMicrotask'),
+      );
+      const macrotaskTasks = executionOrder.filter(
+        (task) =>
+          task.type.includes('setImmediate') ||
+          task.type.includes('setTimeout'),
+      );
+
+      if (microtaskTasks.length > 0 && macrotaskTasks.length > 0) {
+        const lastMicrotask = Math.max(
+          ...microtaskTasks.map((t) => t.timestamp),
+        );
+        const firstMacrotask = Math.min(
+          ...macrotaskTasks.map((t) => t.timestamp),
+        );
+        expect(lastMicrotask).toBeLessThanOrEqual(firstMacrotask);
+      }
+    });
+  });
+});
